+++ lib/aliyun_oss/bucket/website.ex
@@ -5,10 +5,7 @@ defmodule Aliyun.Oss.Bucket.Website do
 
   import Aliyun.Oss.Bucket, only: [get_bucket: 3, put_bucket: 4, delete_bucket: 3]
   alias Aliyun.Oss.Config
-  alias Aliyun.Oss.Client.{Response, Error}
-
-  @type error() ::
-          %Error{body: String.t(), status_code: integer(), parsed_details: map()} | atom()
+  alias Aliyun.Oss.Client.Response
 
   @doc """
   GetBucketWebsite - gets the static website hosting status and redirection rules configured for a bucket.
@@ -36,8 +33,10 @@ defmodule Aliyun.Oss.Bucket.Website do
       iex> Aliyun.Oss.Bucket.Website.get(config, "unkown-bucket")
       {:error,
         %Aliyun.Oss.Client.Error{
-          status_code: 404,
-          parsed_details: %{
+          status: 404,
+          code: "NoSuchBucket",
+          message: "The specified bucket does not exist.",
+          details: %{
             "BucketName" => "unkown-bucket",
             "Code" => "NoSuchBucket",
             "EC" => "0015-00000101",
@@ -45,13 +44,12 @@ defmodule Aliyun.Oss.Bucket.Website do
             "Message" => "The specified bucket does not exist.",
             "RecommendDoc" => "https://api.aliyun.com/troubleshoot?q=0015-00000101",
             "RequestId" => "5C0000000000000000000000"
-          },
-          body: "<?xml version=\"1.0\" encoding=\"UTF-8\"?>...</xml>"
+          }
         }
       }
 
   """
-  @spec get(Config.t(), String.t()) :: {:error, error()} | {:ok, Response.t()}
+  @spec get(Config.t(), String.t()) :: {:error, Exception.t()} | {:ok, Response.t()}
   def get(config, bucket) do
     get_bucket(config, bucket, query_params: %{"website" => nil})
   end
@@ -87,7 +85,7 @@ defmodule Aliyun.Oss.Bucket.Website do
 
   """
   @spec put(Config.t(), String.t(), String.t()) ::
-          {:error, error()} | {:ok, Aliyun.Oss.Client.Response.t()}
+          {:error, Exception.t()} | {:ok, Aliyun.Oss.Client.Response.t()}
   def put(config, bucket, xml_body) do
     put_bucket(config, bucket, xml_body, query_params: %{"website" => nil})
   end
@@ -113,7 +111,10 @@ defmodule Aliyun.Oss.Bucket.Website do
       iex> Aliyun.Oss.Bucket.Website.delete(config, "unknown-bucket")
       {:error,
       %Aliyun.Oss.Client.Error{
-        parsed_details: %{
+        status: 404,
+        code: "NoSuchBucket",
+        message: "The specified bucket does not exist.",
+        details: %{
           "BucketName" => "unknown-bucket",
           "Code" => "NoSuchBucket",
           "EC" => "0015-00000101",
@@ -121,14 +122,12 @@ defmodule Aliyun.Oss.Bucket.Website do
           "Message" => "The specified bucket does not exist.",
           "RecommendDoc" => "https://api.aliyun.com/troubleshoot?q=0015-00000101",
           "RequestId" => "5C38283EC84D1C4471F2F48A"
-        },
-        body: "<?xml version=\"1.0\" encoding=\"UTF-8\"?>...</xml>",
-        status_code: 404
+        }
       }}
 
   """
   @spec delete(Config.t(), String.t()) ::
-          {:error, error()} | {:ok, Aliyun.Oss.Client.Response.t()}
+          {:error, Exception.t()} | {:ok, Aliyun.Oss.Client.Response.t()}
   def delete(config, bucket) do
     delete_bucket(config, bucket, query_params: %{"website" => nil})
   end