+++ lib/aliyun_oss/object/tagging.ex
@@ -5,10 +5,7 @@ defmodule Aliyun.Oss.Object.Tagging do
 
   import Aliyun.Oss.Object, only: [get_object: 4, put_object: 5, delete_object: 4]
   alias Aliyun.Oss.Config
-  alias Aliyun.Oss.Client.{Response, Error}
-
-  @type error() ::
-          %Error{body: String.t(), status_code: integer(), parsed_details: map()} | atom()
+  alias Aliyun.Oss.Client.Response
 
   @doc """
   GetObjectTagging - gets the tags of an object.
@@ -31,7 +28,7 @@ defmodule Aliyun.Oss.Object.Tagging do
       }
 
   """
-  @spec get(Config.t(), String.t(), String.t()) :: {:error, error()} | {:ok, Response.t()}
+  @spec get(Config.t(), String.t(), String.t()) :: {:error, Exception.t()} | {:ok, Response.t()}
   def get(config, bucket, object) do
     get_object(config, bucket, object, headers: %{}, query_params: %{"tagging" => nil})
   end
@@ -66,7 +63,7 @@ defmodule Aliyun.Oss.Object.Tagging do
   </Tagging>
   """
   @spec put(Config.t(), String.t(), String.t(), keyword() | [{String.t(), String.t()}, ...]) ::
-          {:error, error()} | {:ok, Response.t()}
+          {:error, Exception.t()} | {:ok, Response.t()}
   def put(config, bucket, object, tags) when is_list(tags) do
     xml_body = EEx.eval_string(@body_tmpl, tags: tags)
     put_object(config, bucket, object, xml_body, query_params: %{"tagging" => nil})
@@ -88,7 +85,8 @@ defmodule Aliyun.Oss.Object.Tagging do
       }
 
   """
-  @spec delete(Config.t(), String.t(), String.t()) :: {:error, error()} | {:ok, Response.t()}
+  @spec delete(Config.t(), String.t(), String.t()) ::
+          {:error, Exception.t()} | {:ok, Response.t()}
   def delete(config, bucket, object) do
     delete_object(config, bucket, object, query_params: %{"tagging" => nil})
   end